requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
alembic>=1.12.0